Lose AP, back stabs & critical hits, chance to reopen cuts & worsen injuries, bandage MedicalSupply item that speeds up healing & soils over time

+++ Assets/Scripts/Inventory/ItemData.cs
@@ -34,7 +34,7 @@ public class ItemData : MonoBehaviour
     public float blockChanceMultiplier;
 
     [Header("Consumable Data")]
-    public int freshness = 100;
+    public float freshness = 100f;
     public int uses = 1;
     
     [HideInInspector] public Inventory bagInventory;
